+++ client/components/cart.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {connect} from 'react-redux';
-import {fetchCart} from '../store/cart';
+import {fetchCart} from '../store/redux/cart';
 
 class Cart extends React.Component {
   constructor() {
@@ -9,17 +9,13 @@ class Cart extends React.Component {
     this.handleCheckout = this.handleCheckout.bind(this);
   }
 
-  //   do we need async await?
-  async componentDidMount() {
-    await this.props.loadCart(this.props.match.params.cartId);
+  componentDidMount() {
+    this.props.loadCart(this.props.match.params.cartId);
   }
 
-  //   add async when have api route figured out
   handleCheckout() {
     // update quantity (inventory) in db
     // empty cart
-    // redirect to confirmation page
-    // await axios.put(`/api/projects/${this.props.project.id}`, { unassignId });
     console.log('button works');
     window.location.href = '/confirmation';
   }
@@ -41,16 +37,18 @@ class Cart extends React.Component {
     return (
       <div>
         <h1>Shopping Cart</h1>
-        {/* If cart is empty, show 'Cart is emtpy' message, otherwise, show what's below */}
-        Total Items: {totalItems}
-        <br />
-        Order Total: {orderTotal}
-        <br />
-        {/* map over items in cart and show name, image, quantity (with ability to add/subtract), item price and total price */}
-        <br />
-        <button type="submit" onClick={this.handleCheckout}>
-          Proceed to Checkout
-        </button>
+        {cart.length ? (
+          <div>
+            <p>Total Items: {totalItems}</p>
+            <p>Order Total: {orderTotal}</p>
+            {/* map over items in cart and show name, image, quantity (with ability to add/subtract), item price and total price */}
+            <button type="submit" onClick={this.handleCheckout}>
+              Proceed to Checkout
+            </button>
+          </div>
+        ) : (
+          <div>No items in cart. Continue shopping!</div>
+        )}
       </div>
     );
   }
