+++ client/components/index.js
@@ -11,4 +11,5 @@ export {default as AllProducts} from './products/AllProducts';
 export {default as SingleProduct} from './products/SingleProduct';
 export {default as AllUsers} from './users/UserList';
 export {default as SingleUser} from './users/SingleUser';
-//export {default as Cart} from './cart'
+export {default as Cart} from './Cart';
+export {default as Confirmation} from './Confirmation';
