+++ client/index.js
@@ -1,13 +1,13 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import {Provider} from 'react-redux'
-import {Router} from 'react-router-dom'
-import history from './history'
-import store from './store'
-import App from './app'
+import React from 'react';
+import ReactDOM from 'react-dom';
+import {Provider} from 'react-redux';
+import {Router} from 'react-router-dom';
+import history from './history';
+import store from './store';
+import App from './app';
 
 // establishes socket connection
-import './socket'
+import './socket';
 
 ReactDOM.render(
   <Provider store={store}>
@@ -16,4 +16,4 @@ ReactDOM.render(
     </Router>
   </Provider>,
   document.getElementById('app')
-)
+);
