+++ server/api/cart.js
@@ -2,7 +2,7 @@ const router = require('express').Router();
 const Sequelize = require('sequelize');
 const Product = require('../db/models/product');
 const Cart = require('../db/models/cart');
-// const Order = require('../db/models/order');
+const Order = require('../db/models/order');
 
 router.get('/', async (req, res, next) => {
   try {
@@ -17,30 +17,13 @@ router.get('/', async (req, res, next) => {
 
 router.get('/:cartId', async (req, res, next) => {
   try {
-    // look up eager loading for many to many relationship
-    // look into model class and instance methods?
-
     const thisCart = await Cart.findAll({
       where: {
-        // should be able to chagne to findByPk because Hillary added PK to cart table
-        orderId: req.params.cartId
+        order_id: req.params.cartId
       }
-      //   include: {
-      //     model: Product,
-      //     // required: true,
-      //     // where: {
-      //     //   id: Sequelize.col('cart.productId'),
-      //     // },
-      //     through: {
-      //       where: {
-      //         id: Sequelize.col('cart.productId'),
-      //       },
-      //     },
-      //   },
-      //   include Product model where productId: id or id: productId? want to load product info
     });
     if (!thisCart) res.sendStatus(404);
-    res.send(thisCart);
+    res.json(thisCart);
   } catch (error) {
     next(error);
   }
